# Conflicts:
#	src/Orchard.Web/Core/Contents/Views/Parts.Contents.Publish.SummaryAdmin.cshtml
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Drivers/AudioPartDriver.cs
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Drivers/DocumentPartDriver.cs
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Drivers/MediaPartDriver.cs
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Drivers/OEmbedPartDriver.cs

+++ src/Orchard.Web/Modules/Orchard.MediaLibrary/Controllers/ClientStorageController.cs
@@ -1,43 +1,50 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Web.Mvc;
 using Orchard.ContentManagement;
+using Orchard.ContentManagement.Handlers;
+using Orchard.FileSystems.Media;
+using Orchard.Localization;
+using Orchard.Logging;
+using Orchard.MediaLibrary.Models;
 using Orchard.MediaLibrary.Services;
 using Orchard.MediaLibrary.ViewModels;
 using Orchard.Themes;
 using Orchard.UI.Admin;
-using Orchard.MediaLibrary.Models;
-using Orchard.Localization;
-using System.Linq;
-using Orchard.FileSystems.Media;
 
 namespace Orchard.MediaLibrary.Controllers {
     [Admin, Themed(false)]
     public class ClientStorageController : Controller {
         private readonly IMediaLibraryService _mediaLibraryService;
         private readonly IMimeTypeProvider _mimeTypeProvider;
+        private readonly IEnumerable<IContentHandler> _handlers;
 
         public ClientStorageController(
             IMediaLibraryService mediaManagerService,
             IOrchardServices orchardServices,
-            IMimeTypeProvider mimeTypeProvider) {
+            IMimeTypeProvider mimeTypeProvider,
+            IEnumerable<IContentHandler> handlers) {
             _mediaLibraryService = mediaManagerService;
             _mimeTypeProvider = mimeTypeProvider;
+            _handlers = handlers;
             Services = orchardServices;
             T = NullLocalizer.Instance;
+            Logger = NullLogger.Instance;
         }
 
         public IOrchardServices Services { get; set; }
         public Localizer T { get; set; }
+        public ILogger Logger { get; set; }
 
         public ActionResult Index(string folderPath, string type, int? replaceId = null) {
-            if (!Services.Authorizer.Authorize(Permissions.ManageOwnMedia)) {
+            if (!_mediaLibraryService.CheckMediaFolderPermission(Permissions.SelectMediaContent, folderPath)) {
                 return new HttpUnauthorizedResult();
             }
 
             // Check permission
-            if (!Services.Authorizer.Authorize(Permissions.ManageMediaContent) && !_mediaLibraryService.CanManageMediaFolder(folderPath)) {
+            if (!_mediaLibraryService.CanManageMediaFolder(folderPath)) {
                 return new HttpUnauthorizedResult();
             }
 
@@ -59,20 +66,17 @@ namespace Orchard.MediaLibrary.Controllers {
 
         [HttpPost]
         public ActionResult Upload(string folderPath, string type) {
-            if (!Services.Authorizer.Authorize(Permissions.ManageOwnMedia)) {
+            if (!_mediaLibraryService.CheckMediaFolderPermission(Permissions.ImportMediaContent, folderPath)) {
                 return new HttpUnauthorizedResult();
             }
 
             // Check permission
-            if (!Services.Authorizer.Authorize(Permissions.ManageMediaContent) && !_mediaLibraryService.CanManageMediaFolder(folderPath)) {
+            if (!_mediaLibraryService.CanManageMediaFolder(folderPath)) {
                 return new HttpUnauthorizedResult();
             }
 
             var statuses = new List<object>();
             var settings = Services.WorkContext.CurrentSite.As<MediaLibrarySettingsPart>();
-            var allowedExtensions = (settings.UploadAllowedFileTypeWhitelist ?? "")
-                .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
-                .Where(x => x.StartsWith("."));
 
             // Loop through each file in the request
             for (int i = 0; i < HttpContext.Request.Files.Count; i++) {
@@ -86,14 +90,12 @@ namespace Orchard.MediaLibrary.Controllers {
                 }
 
                 // skip file if the allowed extensions is defined and doesn't match
-                if (allowedExtensions.Any()) {
-                    if (!allowedExtensions.Any(e => filename.EndsWith(e, StringComparison.OrdinalIgnoreCase))) {
-                        statuses.Add(new {
-                            error = T("This file type is not allowed: {0}", Path.GetExtension(filename)).Text,
-                            progress = 1.0,
-                        });
-                        continue;
-                    }
+                if (!settings.IsFileAllowed(filename)) {
+                    statuses.Add(new {
+                        error = T("This file is not allowed: {0}", filename).Text,
+                        progress = 1.0,
+                    });
+                    continue;
                 }
 
                 try {
@@ -110,6 +112,7 @@ namespace Orchard.MediaLibrary.Controllers {
                     });
                 }
                 catch (Exception ex) {
+                    Logger.Error(ex, "Unexpected exception when uploading a media.");
                     statuses.Add(new {
                         error = T(ex.Message).Text,
                         progress = 1.0,
@@ -131,17 +134,15 @@ namespace Orchard.MediaLibrary.Controllers {
                 return HttpNotFound();
 
             // Check permission
-            if (!Services.Authorizer.Authorize(Permissions.ManageMediaContent) && !_mediaLibraryService.CanManageMediaFolder(replaceMedia.FolderPath)) {
+            if (!(_mediaLibraryService.CheckMediaFolderPermission(Permissions.EditMediaContent, replaceMedia.FolderPath) && _mediaLibraryService.CheckMediaFolderPermission(Permissions.ImportMediaContent, replaceMedia.FolderPath)) 
+                && !_mediaLibraryService.CanManageMediaFolder(replaceMedia.FolderPath)) {
                 return new HttpUnauthorizedResult();
             }
 
             var statuses = new List<object>();
 
             var settings = Services.WorkContext.CurrentSite.As<MediaLibrarySettingsPart>();
-            var allowedExtensions = (settings.UploadAllowedFileTypeWhitelist ?? "")
-                .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
-                .Where(x => x.StartsWith("."));
-
+            
             // Loop through each file in the request
             for (int i = 0; i < HttpContext.Request.Files.Count; i++) {
                 // Pointer to file
@@ -154,14 +155,12 @@ namespace Orchard.MediaLibrary.Controllers {
                 }
 
                 // skip file if the allowed extensions is defined and doesn't match
-                if (allowedExtensions.Any()) {
-                    if (!allowedExtensions.Any(e => filename.EndsWith(e, StringComparison.OrdinalIgnoreCase))) {
-                        statuses.Add(new {
-                            error = T("This file type is not allowed: {0}", Path.GetExtension(filename)).Text,
-                            progress = 1.0,
-                        });
-                        continue;
-                    }
+                if (!settings.IsFileAllowed(filename)) {
+                    statuses.Add(new {
+                        error = T("This file is not allowed: {0}", filename).Text,
+                        progress = 1.0,
+                    });
+                    continue;
                 }
 
                 try {
@@ -171,13 +170,29 @@ namespace Orchard.MediaLibrary.Controllers {
                     if (!replaceContentType.Equals(replaceMedia.TypeDefinition.Name, StringComparison.OrdinalIgnoreCase))
                         throw new Exception(T("Cannot replace {0} with {1}", replaceMedia.TypeDefinition.Name, replaceContentType).Text);
 
-                    _mediaLibraryService.DeleteFile(replaceMedia.FolderPath, replaceMedia.FileName);
+                    // Raise update and publish events which will update relevant Media properties
+                    _handlers.Invoke(x => x.Updating(new UpdateContentContext(replaceMedia.ContentItem)), Logger);
+
+                    var mediaItemsUsingTheFile = Services.ContentManager.Query<MediaPart, MediaPartRecord>()
+                                                                .ForVersion(VersionOptions.Latest)
+                                                                .Where(x => x.FolderPath == replaceMedia.FolderPath && x.FileName == replaceMedia.FileName)
+                                                                .Count();
+                    if (mediaItemsUsingTheFile == 1) { // if the file is referenced only by the deleted media content, the file too can be removed.
+                        _mediaLibraryService.DeleteFile(replaceMedia.FolderPath, replaceMedia.FileName);
+                    }
+                    else {
+                        // it changes the media file name
+                        replaceMedia.FileName = filename;
+                    }
                     _mediaLibraryService.UploadMediaFile(replaceMedia.FolderPath, replaceMedia.FileName, file.InputStream);
                     replaceMedia.MimeType = mimeType;
 
-                    // Force a publish event which will update relevant Media properties
-                    replaceMedia.ContentItem.VersionRecord.Published = false;
-                    Services.ContentManager.Publish(replaceMedia.ContentItem);
+                    _handlers.Invoke(x => x.Updated(new UpdateContentContext(replaceMedia.ContentItem)), Logger);
+
+                    var publishedVersion = replaceMedia.ContentItem.Record.Versions.SingleOrDefault(v => v.Published);
+
+                    _handlers.Invoke(x => x.Publishing(new PublishContentContext(replaceMedia.ContentItem, publishedVersion)), Logger);
+                    _handlers.Invoke(x => x.Published(new PublishContentContext(replaceMedia.ContentItem, publishedVersion)), Logger);
 
                     statuses.Add(new {
                         id = replaceMedia.Id,
@@ -189,6 +204,8 @@ namespace Orchard.MediaLibrary.Controllers {
                     });
                 }
                 catch (Exception ex) {
+                    Logger.Error(ex, "Unexpected exception when uploading a media.");
+
                     statuses.Add(new {
                         error = T(ex.Message).Text,
                         progress = 1.0,
